🧬 Modify main reset section
I prefer CSS resets to be concise and elegant. Excluding a number of
elements simply to maintain their dimensional attributes is neither
concise nor elegant. Therefore, I removed iframe, canvas, img, svg,
video, and audio from the exclusion list, because I believe that the
dimensions of these elements will usually be set directly in CSS or
based on their parent containers anyways.

I moved the list-style reset to this block as I find it more concise and
elegant that having yet another selector that reset styles.

I exclude the outline property because I place a high value on
accessibility, and I do not think that having to restyle focus in
every style sheet is elegant for something that is rarely seen or styled
otherwise.