🧬 Multiple changes to make concise and elegant
I have added where to all the selectors so that the reset remains at 0
specificity.

As part of the desire to be concise rather than simple revert all for
various fixes, I have found the actual issue and reverted only those
properties.  I find this much more elegant and DRY (as demonstrated by
the need to add a block sizing fix to the original recently.)

Tables are rarely used anymore and when they are I feel that adding the
border-collapse setting is trivial at that point.

I removed the fix for the case of setting user-select:none on the body
tag as I find that practice to be terrible in most cases, and I am not
going to enable it. In the rare case this is an appropriate action,
the `fix` can be implemented.